split Map + Overlay management out of <App>
In my haste to get the scrolling bottom drawer working a while back, I
put a bunch of crap into the App component that handled scrolling,
positioning the map controls, forwarding touch events and stuff like
that.

This was all mixed in with the top-level logic of what content to
render in the top bar and bottom drawer, depending on global
application state.

This commit splits that out. <App> decides what content to render at
the top and bottom based on app state, while the new <MapPlusOverlay>
handles the mechanics of the map overlay, scrolling, touch events,
etc.

In case it's not obvious, all of the code in MapPlusOverlay here is
not new code, but is moved from App with no changes except to rename
CSS classes.
