+++ src/components/MapPlusOverlay.css
@@ -1,0 +1,74 @@
+.MapPlusOverlay {
+  display: flex;
+  flex-flow: column nowrap;
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+}
+
+.MapPlusOverlay_column {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  z-index: 10;
+  flex-shrink: 1;
+  min-height: 10px;
+}
+
+.MapPlusOverlay_column__nonTouchDevice {
+  pointer-events: none;
+}
+
+.MapPlusOverlay_column__scrollable {
+  pointer-events: auto;
+}
+
+.MapPlusOverlay_overlay {
+  display: flex;
+  flex-grow: 1;
+  flex-direction: column;
+  overflow-y: scroll;
+}
+
+.MapPlusOverlay_bottomPane {
+  background: white;
+  border-radius: 9px 9px 0 0;
+  pointer-events: auto;
+  box-shadow: 0 1px 10px #c4c4c4;
+
+  /* Make sure you can always scroll down even when iOS, etc. is putting an
+   * address bar over the bottom of the page.
+   */
+  padding-bottom: 0; /* For browsers not supporting env */
+  padding-bottom: env(safe-area-inset-bottom);
+  /* See below as well */
+}
+
+body.isIOSChromeOrSafari .MapPlusOverlay_bottomPane {
+  /* Apple documentation suggests you can rely on env(safe-area-inset-bottom)
+   * here, but it's 0 on iOS 15 on my iPhone SE 1st gen, despite the expanding
+   * address bar. Apparently can't be relied upon. The fallback value below is
+   * empirically determined and applies to Chrome and Safari on iOS but is not
+   * needed for some reason on Firefox for iOS
+   */
+  padding-bottom: max(env(safe-area-inset-bottom), 74px);
+}
+
+@media all and (display-mode: standalone) {
+  body.isIOSChromeOrSafari .MapPlusOverlay_bottomPane {
+    /* When running as a PWA, don't add extra padding */
+    padding-bottom: env(safe-area-inset-bottom);
+  }
+}
+
+.MapPlusOverlay_bottomPane__withMapHidden {
+  flex-grow: 1;
+}
+
+/* The transparent section of the overlay, through which the map can
+ * be viewed and interacted with. */
+.MapPlusOverlay_window {
+  flex-basis: calc(100% - 140px);
+  flex-shrink: 0;
+  flex-grow: 1;
+}
